requirements.txt
streamlit==1.20.0
plotly==5.13.1
pandas==1.4.0
numpy==1.22.1
folium==0.14.0
matplotlib==3.5.1
matplotlib_inline==0.1.2
haversine==2.8.0
streamlit_folium==0.11.1
Pillow==7.0.0